Moved `format_expert_conversation` to shared utilities, updated relevant imports, and removed redundant function definitions from RAG modules. Simplified conversation formatting logic in ExpertRAGAgent.
